% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_flowjo.R
\name{import_flowjo}
\alias{import_flowjo}
\title{Importing raw .csv files directly from FlowJo}
\usage{
import_flowjo(
  directory = "DataRaw/",
  filter_string = "\\\\.csv",
  rm.bottom = T
)
}
\arguments{
\item{directory}{a string with the path to the folder containng csv files}

\item{filter_string}{a string that has a pattern filter out csv files}

\item{rm.bottomm}{T if std and mean are at the bottomm and to remove the last column. F if you just want to import and concatenate everything}
}
\value{

}
\description{
Imports csv files created directly from FlowJo and removed Mean, SD, and empty final column. In addition if there are NAs in the table the sample is removed at this step too!
The defaults take all the csv in the DataRaw/ folder, but changing the filter_string can allow you to be selective about which files to merge
}
\author{
akirosingh
}
