% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_cat.R
\name{wilcox_cat}
\alias{wilcox_cat}
\title{}
\usage{
wilcox_cat(df, x_cat, y_var, condition, paired = F)
}
\value{

}
\description{
.. content for \description{Wrapper-function to do a wilcox test across multiple conditions (Think Stim conditions) } (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
akirosingh
}
