% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_samples.R
\name{intersect_samples}
\alias{intersect_samples}
\title{Core samples across conditions}
\usage{
intersect_samples(df, sampleid, condition = NA, x_cat)
}
\arguments{
\item{df}{the dataframe to subset to sample ids shared across x_cat}

\item{sampleid}{the sample ids often patient numbers}

\item{condition}{the experimental condition such as Stim condition == "M1"}

\item{x_cat}{the variable which has different sample ids across like days}
}
\value{

}
\description{
Finding the core samples shared across many different conditions for paired tests content for paired tests
}
\author{
akirosingh
}
