% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cat.R
\name{plot_cat}
\alias{plot_cat}
\title{}
\usage{
plot_cat(df, x_cat, y_var, condition, wilcox_results)
}
\value{

}
\description{
.. content for \description{Wrapper function to plot boxplots with wilcox significance added from previous wilcox_cat function.} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
akirosingh
}
