% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_num.R
\name{plot_num}
\alias{plot_num}
\title{}
\usage{
plot_num(df, x_num, y_var, condition, correlation_results)
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
akirosingh
}
