% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_num.R
\name{plot_num}
\alias{plot_num}
\title{Graphing spearman correlation with scatterplot}
\usage{
plot_num(df, x_num, y_var, condition, correlation_results)
}
\arguments{
\item{df}{data}

\item{x_num}{a continuous variable}

\item{y_var}{another continuous variable}

\item{condition}{the different experimental conditions in an experiment}

\item{correlation_results}{results from spearman tests}
}
\value{

}
\description{
.. content for \description{Generating a scatterplot with a spearman correlation.} (no empty lines) ..
}
\author{
akirosingh
}
