% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_flowjo.R
\name{rename_flowjo}
\alias{rename_flowjo}
\title{Fixing the column names of csv files exported from flowjo}
\usage{
rename_flowjo(data, removal = " \\\\| Freq\\\\. of Parent \\\\(\\\\\%\\\\)")
}
\arguments{
\item{data}{dataframe from flowjo with raw column names}

\item{removal}{the parts of the column names to remove}
}
\value{

}
\description{
.. content for \description{Fixes the default column names for raw FlowJo CSV files. It takes out the freq of parent clause at the end of the name and takes out the last two cell features for the name. TODO: Make sure this works with boolean names} (no empty lines) ..
}
\author{
akirosingh
}
